add "updated..." timestamp in page footer
To get space for it, remove the link to the privacy statement. This is
acceptable because the privacy statement is referenced from within the
Terms of Service, for which there is still a link in the footer.
Furthermore, the help site does not have user accounts etc., so giving
the URL to the privacy statement alongside the Terms of Service is not
as well motivated.