add filename-rewriting rule to collect images
This changeset adds the case of stripping the prefix `480px-` if
present, so that the files with and without the prefix are copied
during `make`.

The routine is ready to easily add more rules for rewriting filenames
so as to collect more image files that are otherwise not obviously
found according to Markdown syntax.