if full-size photo missing, try to switch format
The full-size rendering of a photograph is obtained automatically by
removing a prefix like `480px-` if present. To allow for the two image
files to use different formats, in addition to having different sizes,
this cangeset introduces a conditional branch that checks whether a
file with the default ending exists and, if not, switches it to '.jpg'
or '.png', accordingly.