follow handmotion.ipynb on `master` branch
Compared with the commit hash-based URL, this introduces a risk of a
future dead URL, e.g., if the file is moved in the sourcetree.
However, the Python client is in active development, so it is easier
to point to the current tip as the ipynb file is changing frequently.

Furthermore, the URL to Colab is based on the tip of `master` branch.